+++ weblate/trans/models/unit.py
@@ -763,7 +763,7 @@ class Unit(models.Model, LoggerMixin):
                 )
 
     def update_source_unit(
-        self, component, source, context, pos, note, location, flags, explanation
+        self, component, source, context, pos, note, location, flags: Flags, explanation
     ) -> None:
         source_unit = component.get_source(
             self.id_hash,
@@ -775,10 +775,10 @@ class Unit(models.Model, LoggerMixin):
                 "note": note,
                 "location": location,
                 "explanation": explanation,
-                "flags": flags,
+                "flags": flags.format(),
             },
         )
-        same_flags = flags == source_unit.flags
+        same_flags = flags == Flags(source_unit.flags)
         if (
             not source_unit.source_updated
             and not source_unit.translation.filename
@@ -793,7 +793,7 @@ class Unit(models.Model, LoggerMixin):
             source_unit.source_updated = True
             source_unit.location = location
             source_unit.explanation = explanation
-            source_unit.flags = flags
+            source_unit.flags = flags.format()
             source_unit.note = note
             source_unit.save(
                 update_fields=["position", "location", "explanation", "flags", "note"],
@@ -803,7 +803,15 @@ class Unit(models.Model, LoggerMixin):
             )
         self.source_unit = source_unit
 
-    def update_from_unit(self, unit, pos, created) -> None:  # noqa: C901
+    def update_from_unit(  # noqa: C901
+        self,
+        *,
+        unit: TranslationUnit,
+        pos: int,
+        created: bool,
+        user: User | None = None,
+        author: User | None = None,
+    ) -> None:
         """Update Unit from ttkit unit."""
         translation = self.translation
         component = translation.component
@@ -819,7 +827,7 @@ class Unit(models.Model, LoggerMixin):
             else:
                 explanation = self.explanation
                 source_explanation = "" if created else self.source_unit.explanation
-            flags = unit.flags
+            flags = Flags(unit.flags)
             source = unit.source
             self.check_valid(split_plural(source))
             if not translation.is_template and translation.is_source:
@@ -898,7 +906,7 @@ class Unit(models.Model, LoggerMixin):
                 previous_source = self.previous_source
 
         # Update checks on fuzzy update or on content change
-        same_state = state == self.state and flags == self.flags
+        same_state = state == self.state and flags == Flags(self.flags)
         same_metadata = (
             location == self.location
             and explanation == self.explanation
@@ -912,7 +920,7 @@ class Unit(models.Model, LoggerMixin):
             and same_target
             and same_state
             and original_state == self.original_state
-            and flags == self.flags
+            and flags == Flags(self.flags)
             and previous_source == self.previous_source
             and self.source_unit == old_source_unit
             and old_source_unit is not None
@@ -927,7 +935,7 @@ class Unit(models.Model, LoggerMixin):
         self.position = pos
         self.location = location
         self.explanation = explanation
-        self.flags = flags
+        self.flags = flags.format()
         self.source = source
         self.target = target
         self.state = state
@@ -968,8 +976,8 @@ class Unit(models.Model, LoggerMixin):
         if not same_source and source_change:
             translation.update_changes.append(
                 self.generate_change(
-                    None,
-                    None,
+                    user,
+                    author,
                     ActionEvents.SOURCE_CHANGE,
                     check_new=False,
                     old=source_change,
@@ -981,8 +989,8 @@ class Unit(models.Model, LoggerMixin):
         if not same_data:
             translation.update_changes.append(
                 self.generate_change(
-                    user=None,
-                    author=None,
+                    user,
+                    author,
                     change_action=translation.create_unit_change_action
                     if created
                     else translation.update_unit_change_action,
@@ -2000,10 +2008,16 @@ class Unit(models.Model, LoggerMixin):
                     unit=unit,
                     author=user,
                 )
-            PendingUnitChange.store_unit_change(
-                unit=self,
-                author=user,
-            )
+
+            # translation file does not exist for source strings in bilingual formats
+            if (
+                not self.is_source
+                or self.translation.component.file_format_cls.monolingual
+            ):
+                PendingUnitChange.store_unit_change(
+                    unit=self,
+                    author=user,
+                )
 
         if save:
             self.save(update_fields=["explanation"], only_save=True)