+++ weblate/trans/tests/test_models.py
@@ -6,6 +6,7 @@
 
 import os
 from datetime import timedelta
+from unittest.mock import patch
 
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from django.core.management.color import no_style
@@ -16,6 +17,7 @@ from django.utils import timezone
 
 from weblate.auth.models import Group, User
 from weblate.checks.models import Check
+from weblate.formats.base import UnitNotFoundError
 from weblate.lang.models import Language, Plural
 from weblate.trans.actions import ActionEvents
 from weblate.trans.exceptions import SuggestionSimilarToTranslationError
@@ -287,6 +289,15 @@ class TranslationTest(RepoTestCase):
         self.assertEqual(start_rev, component.repository.last_revision)
         self.assertEqual(translation.count_pending_units, 4)
 
+        with patch.object(
+            translation.store, "find_unit", side_effect=UnitNotFoundError("test error")
+        ):
+            # Check failure in committing pending changes retains them in db
+            # use _commit_pending, patching won't work if the Translation object
+            # is loaded from the db
+            translation._commit_pending("test", None)  # noqa: SLF001
+            self.assertEqual(translation.count_pending_units, 4)
+
         # Commit pending changes
         translation.commit_pending("test", None)
         self.assertNotEqual(start_rev, component.repository.last_revision)
@@ -329,6 +340,32 @@ class TranslationTest(RepoTestCase):
         self.assertEqual(author, user2)
         self.assertEqual(len(changes), 1)
         self.assertEqual([c.unit for c in changes], [units[3]])
+
+    def test_commit_explanation(self):
+        user = create_test_user()
+        component = self.create_tbx()
+
+        self.assertEqual(PendingUnitChange.objects.count(), 0)
+
+        en_translation = component.translation_set.get(language_code="en")
+        source_unit = en_translation.unit_set.get(source="address bar")
+        source_unit.update_explanation("explanation 1", user)
+
+        # one change for each translation file, source language does not have
+        # a translation file
+        self.assertFalse(PendingUnitChange.objects.filter(unit=source_unit).exists())
+        self.assertEqual(
+            PendingUnitChange.objects.count(), component.translation_set.count() - 1
+        )
+        component.commit_pending("test", None)
+
+        self.assertEqual(PendingUnitChange.objects.count(), 0)
+
+        cs_translation = component.translation_set.get(language_code="cs")
+        target_unit = cs_translation.unit_set.get(source="address bar")
+        target_unit.update_explanation("explanation 2", user)
+        # only adds pending change for target unit's translation file
+        self.assertEqual(PendingUnitChange.objects.count(), 1)
 
 
 class ComponentListTest(RepoTestCase):